separated pages and components files
