--- src/components/utils/validators/index.ts
@@ -1,1 +1,0 @@
-export { maxTextLength, required, composeValidators } from './validators';
