+++ src/pages/profile/Profile.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { ProfileType } from '../../@types/types.ts';
 import MyPostsContainer from './myPosts/MyPostsContainer.tsx';
 import st from './Profile.module.css';
-import { ProfileInfo } from './profileInfo';
+import { ProfileInfo } from './profileInfo/index.ts';
 
 type Props = {
     profile: ProfileType | null;
